services/api/requirements.txt
# ─── FastAPI Core ─────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ─── Database / ORM ───────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0            # PostgreSQL async driver
alembic==1.13.1            # Database migrations

# ─── Pydantic / Settings ──────────────────────────────────────────────────────
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.0

# ─── Auth / JWT ───────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.27.0              # Async HTTP (Hacienda API, OAuth2)
tenacity==8.2.3            # Retry logic

# ─── Redis / Celery (Workers) ─────────────────────────────────────────────────
celery==5.3.6
redis==5.0.3
kombu==5.3.4

# ─── Digital Signing / Cryptography ──────────────────────────────────────────
cryptography==42.0.5       # .p12 loading, RSA-SHA256 signature (XAdES-BES)
pyOpenSSL==24.0.0          # Complemento para OpenSSL operations

# ─── XML Generation ───────────────────────────────────────────────────────────
lxml==5.1.0                # Alternativa más rápida al xml.etree estándar

# ─── PDF Generation ───────────────────────────────────────────────────────────
reportlab==4.1.0           # Layout PDF profesional
qrcode[pil]==7.4.2         # QR codes para URL verificación Hacienda
Pillow==10.2.0             # Procesamiento de imágenes

# ─── CABYS / Public Hacienda API ──────────────────────────────────────────────
cachetools==5.3.3          # Cache de búsquedas CABYS y tipo de cambio

# ─── Utils ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1       # .env loading
structlog==24.1.0          # Structured logging
orjson==3.9.15             # Fast JSON serialization
anyio==4.3.0               # Async compatibility
